Fixed models folder structure for merge

+++ App/models/user.py
@@ -4,7 +4,7 @@ from werkzeug.security import check_password_hash, generate_password_hash
 class User(db.Model):
     __tablename__ = 'user'
 
-    u_ID = db.Column(db.Integer, primary_key=True)
+    u_ID = db.Column(db.Integer(8), primary_key=True)
     password = db.Column(db.String(120), nullable=False)
 
     def __init__(self, u_ID, password, role):
